requirements.txt
flask
google-generativeai
ipython
ipykernel
python-dotenv